refactor: controller 책임 분배
